# Refactor locs, dependencies into generic sub build