--- a/cmake/API.cmake
+++ b/cmake/API.cmake
@@ -515,15 +515,17 @@ endmacro(register_fprime_ut_target)
 # Helper function to do the actual registration. Also used to side-load prescan to bypass the not-on-prescan check.
 ####
 macro(register_fprime_list_helper TARGET_FILE_PATH TARGET_LIST)
-    include("${TARGET_FILE_PATH}")
-    # Prevent out-of-order setups
-    get_property(MODULE_DETECTION_STARTED GLOBAL PROPERTY MODULE_DETECTION SET)
-    if (MODULE_DETECTION_STARTED)
-        message(FATAL_ERROR "Cannot register fprime target after including subdirectories or FPrime-Code.cmake'")
-    endif()
-    get_property(TARGETS GLOBAL PROPERTY "${TARGET_LIST}")
-    if (NOT TARGET_FILE_PATH IN_LIST TARGETS)
-        set_property(GLOBAL APPEND PROPERTY "${TARGET_LIST}" "${TARGET_FILE_PATH}")
+    if (NOT DEFINED FPRIME_SUB_BUILD_TARGETS OR "${TARGET_FILE_PATH}" IN_LIST FPRIME_SUB_BUILD_TARGETS)
+        include("${TARGET_FILE_PATH}")
+        # Prevent out-of-order setups
+        get_property(MODULE_DETECTION_STARTED GLOBAL PROPERTY MODULE_DETECTION SET)
+        if (MODULE_DETECTION_STARTED)
+            message(FATAL_ERROR "Cannot register fprime target after including subdirectories or FPrime-Code.cmake'")
+        endif()
+        get_property(TARGETS GLOBAL PROPERTY "${TARGET_LIST}")
+        if (NOT TARGET_FILE_PATH IN_LIST TARGETS)
+            set_property(GLOBAL APPEND PROPERTY "${TARGET_LIST}" "${TARGET_FILE_PATH}")
+        endif()
     endif()
 endmacro(register_fprime_list_helper)