--- a/cmake/API.cmake
+++ b/cmake/API.cmake
@@ -11,6 +11,7 @@
 # - Register an fprime build target/build stage to allow custom build steps. (Experimental)
 #
 ####
+include_guard()
 set(FPRIME_TARGET_LIST "" CACHE INTERNAL "FPRIME_TARGET_LIST: custom fprime targets" FORCE)
 set(FPRIME_UT_TARGET_LIST "" CACHE INTERNAL "FPRIME_UT_TARGET_LIST: custom fprime targets" FORCE)
 set(FPRIME_AUTOCODER_TARGET_LIST "" CACHE INTERNAL "FPRIME_AUTOCODER_TARGET_LIST: custom fprime targets" FORCE)