--- a/cmake/sub-build/sub-build.cmake
+++ b/cmake/sub-build/sub-build.cmake
@@ -10,6 +10,7 @@
 # maintaining the efficiency of properly expressed builds.
 #
 ####
+include_guard()
 include(utilities)
 include(sub-build/sub-build-config)