| | |
|:---|:---|
|**_Originating Project/Creator_**| |
|**_Affected Component_**|  |
|**_Affected Architectures(s)_**|  |
|**_Related Issue(s)_**|  |
|**_Has Unit Tests (y/n)_**|  |
|**_Builds Without Errors (y/n)_**|  |
|**_Unit Tests Pass (y/n)_**|  |
|**_Documentation Included (y/n)_**|  |

---
## Change Description

Creates generic sub-builds and makes locs/dependencies into a sub-build.  This replaces the hardcoded prescan function.

## Rationale

Sub-builds are a standard pattern in CMake and this exposes this functionality in a generic form to users.  Reworking pre-scan into this has the following advantages:

1. "Rebuilding" loc/dependencies does not take substantial time after the first run
2. Proper dependency tracking and rebuilds work
3. `-j` will work for sub/builds and fpp-depend isn't mapped to the whole CPU.

## Future Work

Note any additional work that will be done relating to this issue.


## Review fixes tracked as check-list:
- [x] Remove commented out FPrime-Code
- [x] Remove comment link to pre-scan
- [ ] Remove Fprime-Code.cmake
- [x] Change [prescan] to [subbuild] or [subbuild-name]
- [x] Always print executing step