@LeStarch 2023-11-07
Note: although structurally much nicer, this slows down the initial generate time.  Here are some numbers:

```
Original (Ref): initial generation
-- Configuring done (14.0s)
-- Generating done (0.6s)

This PR (Ref): initial generation w/ 16 jobs
-- Configuring done (17.7s)
-- Generating done (0.6s)

Original (Ref): re-generation
-- Configuring done (12.7s)
-- Generating done (0.5s)

This PR (Ref): initial re-generation w/ 16 jobs
-- Configuring done (6.0s)
-- Generating done (0.6s)
```

Notice that the re-generation (changed a model/CMakeLists.txt) is substantially faster.